// modules/transfer/modules/bulk-transfer/enum/bte-file-validations.enum.ts
export enum EHeadingFileKeys {
  KEY = 'key',
  SOURCE_ACCOUNT = 'sourceAccount',
  CREDITS = 'credits',
  TOTAL_AMOUNT = 'totalAmount',
  CURRENCY = 'currency',
}

export enum EFileDetailsKeys {
  KEY = 'key',
  IDENTIFIER = 'identifier',
  ACCOUNT = 'account',
  NAME = 'name',
  AMOUNT = 'amount',
}